refactor: Move specification rules to a separate internal module.
